+++ packages/notificare-push-ui/src/internal/ui/notification-presenter.ts
@@ -210,11 +210,11 @@ async function presentPassbook(
   }
 
   if (isAppleDevice() && isSafariBrowser()) {
-    window.location.href = `${options.services.pushHost}/pass/pkpass/${id}`;
+    window.location.href = `https://${options.hosts.restApi}/pass/pkpass/${id}`;
     return;
   }
 
-  window.location.href = `${options.services.pushHost}/pass/web/${id}?showWebVersion=1`;
+  window.location.href = `https://${options.hosts.restApi}/pass/web/${id}?showWebVersion=1`;
 }
 
 async function presentUrlScheme(notification: NotificareNotification) {