+++ packages/notificare-push/src/internal/web-push/configuration/worker-configuration.ts
@@ -1,7 +1,7 @@
 export interface WorkerConfiguration {
+  cloudHost: string;
   applicationKey: string;
   applicationSecret: string;
   deviceId?: string;
-  useTestEnvironment?: boolean;
   standalone?: boolean;
 }